added long int support to match the constraints

+++ src.c
@@ -3,9 +3,9 @@
 #include <stdlib.h>
 #include "mpi.h"
 
-void exchangeGhostLayers(double *ghostLayer, double *localArr, int ip, int jp, int kp, int px, int py, int pz, int snx, int sny, int snz, int direction, int sendFlag)
+void exchangeGhostLayers(double *ghostLayer, double *localArr, long long ip, long long jp, long long kp, long long px, long long py, long long pz, long long snx, long long sny, long long snz, long long direction, long long sendFlag)
 {
-    int currRank = ip + px * jp + px * py * kp, neighbourRank;
+    long long currRank = ip + px * jp + px * py * kp, neighbourRank;
     MPI_Datatype newType;
     MPI_Request request;
 
@@ -126,13 +126,13 @@ int main(int argc, char *argv[])
         return 0;
     }
     char *inputFile = argv[1]; // input file
-    int px = atoi(argv[2]);    // number of processes in each direction
-    int py = atoi(argv[3]);
-    int pz = atoi(argv[4]);
-    int nx = atoi(argv[5]); // number of grid points in each direction
-    int ny = atoi(argv[6]);
-    int nz = atoi(argv[7]);
-    int nc = atoi(argv[8]);     // number of time steps
+    long long px = atoll(argv[2]);    // number of processes in each direction
+    long long py = atoll(argv[3]);
+    long long pz = atoll(argv[4]);
+    long long nx = atoll(argv[5]); // number of grid points in each direction
+    long long ny = atoll(argv[6]);
+    long long nz = atoll(argv[7]);
+    long long nc = atoll(argv[8]);     // number of time steps
     char *outputFile = argv[9]; // output file
 
     // Check if the number of processes is correct
@@ -140,32 +140,32 @@ int main(int argc, char *argv[])
     {
         if (rank == 0)
         {
-            printf("[ERROR] The number of processes must be equal to px * py * pz = %d but were given %d\n", px * py * pz, size);
+            printf("[ERROR] The number of processes must be equal to px * py * pz = %lld but were given %d\n", px * py * pz, size);
         }
         return 0;
     }
 
     // subdomain sizes
-    int snx = nx / px;
-    int sny = ny / py;
-    int snz = nz / pz;
+    long long snx = nx / px;
+    long long sny = ny / py;
+    long long snz = nz / pz;
 
-    int N = nx * ny * nz;
-    int P = px * py * pz;
-    int localN = snx * sny * snz;
+    long long N = nx * ny * nz;
+    long long P = px * py * pz;
+    long long localN = snx * sny * snz;
 
     // also get the current process's coordinates
-    int ip = (rank % (px * py)) % px;
-    int jp = (rank % (px * py)) / px;
-    int kp = (rank / (px * py));
+    long long ip = (rank % (px * py)) % px;
+    long long jp = (rank % (px * py)) / px;
+    long long kp = (rank / (px * py));
 
     // Initializing variables to keep a track of time taken in each part of execution
-    int localMinima[nc], localMaxima[nc], totalLocalMinima[nc], totalLocalMaxima[nc];
+    long long localMinima[nc], localMaxima[nc], totalLocalMinima[nc], totalLocalMaxima[nc];
     double globalMinima[nc], globalMaxima[nc], globalGlobalMinima[nc], globalGlobalMaxima[nc];
     double time_1, time_2, time_3, avg_finalReadTime = 0, avg_finalMainTime = 0, avg_finalTotalTime = 0;
 
     // Looping over the number of iterations to get the average time taken
-    for (int i = 0; i < 2; i++)
+    for (long long i = 0; i < 2; i++)
     {
         time_1 = MPI_Wtime();
 
@@ -174,7 +174,7 @@ int main(int argc, char *argv[])
         // Reading the input file globally in the rank 0 process and then distributing it to all processes
         double *arr[nc];
         if (rank == 0) {
-            for (int t = 0; t < nc; t++){
+            for (long long t = 0; t < nc; t++){
                 arr[t] = malloc(N * sizeof(double));
             }
 
@@ -185,8 +185,8 @@ int main(int argc, char *argv[])
                 return 0;
             }
 
-            int xi, yi, zi, sxi, syi, szi, si, pxi, pyi, pzi, pi, offset, i_mod;
-            for (int i = 0; i < N; i++) {
+            long long xi, yi, zi, sxi, syi, szi, si, pxi, pyi, pzi, pi, offset, i_mod;
+            for (long long i = 0; i < N; i++) {
                 // Get the coordinates in current domain
                 xi = (i % (nx * ny)) % nx;
                 yi = (i % (nx * ny)) / nx;
@@ -208,7 +208,7 @@ int main(int argc, char *argv[])
                 offset = pi * snx * sny * snz;
                 i_mod = si + offset;
 
-                for (int t = 0; t < nc; t++){
+                for (long long t = 0; t < nc; t++){
                     fread(&arr[t][i_mod], sizeof(double), 1, file);
                 }
             }
@@ -218,15 +218,15 @@ int main(int argc, char *argv[])
 
         // DISTRIBUTING STRATEGY - As we have already read the data in specific format we will directly distribute the data using MPI_Scatter
         double *localArr[nc];
-        for (int t = 0; t < nc; t++){
+        for (long long t = 0; t < nc; t++){
             localArr[t] = malloc(localN * sizeof(double));
         }
-        for (int t = 0; t < nc; t++){
+        for (long long t = 0; t < nc; t++){
             MPI_Scatter(arr[t], localN, MPI_DOUBLE, localArr[t], localN, MPI_DOUBLE, 0, MPI_COMM_WORLD);
         }
         printf("[DEBUG] Distributed input data to process %d\n", rank);
         if (rank == 0){
-            for (int t = 0; t < nc; t++){
+            for (long long t = 0; t < nc; t++){
                 free(arr[t]);
             }
         }
@@ -235,24 +235,24 @@ int main(int argc, char *argv[])
         // READING AND DISTRIBUTION STRATEGY 2 - parallel I/O
         /*
         double *localArr[nc];
-        for (int t = 0; t < nc; t++)
+        for (long long t = 0; t < nc; t++)
         {
             localArr[t] = malloc(localN * sizeof(double));
         }
 
         // define starting position of each process
-        int startPos = (ip * snx * nc) + (jp * sny * nx * nc) + (kp * snz * nx * ny * nc);
+        long long startPos = (ip * snx * nc) + (jp * sny * nx * nc) + (kp * snz * nx * ny * nc);
 
         // define number of reads for each process and count of each read
-        int numReads = sny * snz;
-        int readCount = nc * snx;
+        long long numReads = sny * snz;
+        long long readCount = nc * snx;
 
         // start reading the file
         MPI_File fh;
-        int offset = 0;
+        long long offset = 0;
         double *readArr = malloc(readCount * numReads * sizeof(double));
         MPI_File_open(MPI_COMM_WORLD, inputFile, MPI_MODE_RDONLY, MPI_INFO_NULL, &fh);
-        for (int k = 0; k < numReads; k++)
+        for (long long k = 0; k < numReads; k++)
         {
             offset = (k % sny) * nx * nc + (k / sny) * nx * ny * nc;
             MPI_File_read_at(fh, (startPos + offset) * sizeof(double), readArr + k * readCount, readCount, MPI_DOUBLE, MPI_STATUS_IGNORE);
@@ -260,9 +260,9 @@ int main(int argc, char *argv[])
         MPI_File_close(&fh);
 
         // convert the data in required format
-        for (int i = 0; i < localN; i++)
+        for (long long i = 0; i < localN; i++)
         {
-            for (int t = 0; t < nc; t++)
+            for (long long t = 0; t < nc; t++)
             {
                 localArr[t][i] = readArr[i * nc + t];
             }
@@ -273,22 +273,22 @@ int main(int argc, char *argv[])
 
         // READING AND DISTRIBUTION STRATEGY 3 - parallel I/O w/ MPI_Type_vector
         double *localArr[nc];
-        for (int t = 0; t < nc; t++)
+        for (long long t = 0; t < nc; t++)
         {
             localArr[t] = malloc(localN * sizeof(double));
         }
 
         // define starting position of each process
-        int startPos = (ip * snx * nc) + (jp * sny * nx * nc) + (kp * snz * nx * ny * nc);
+        long long startPos = (ip * snx * nc) + (jp * sny * nx * nc) + (kp * snz * nx * ny * nc);
 
         // define number of reads for each process and count of each read
-        int numReads = sny * snz;
-        int readCount = nc * snx;
+        long long numReads = sny * snz;
+        long long readCount = nc * snx;
 
         // define blocklengths and displacements for MPI_Type_vector
         int *blocklengths  = malloc(numReads * sizeof(int));
         int *displacements = malloc(numReads * sizeof(int));
-        for (int k = 0; k < numReads; k++)
+        for (long long k = 0; k < numReads; k++)
         {
             blocklengths[k]  = readCount;
             displacements[k] = ((k % sny) * nx * nc + (k / sny) * nx * ny * nc);
@@ -306,9 +306,9 @@ int main(int argc, char *argv[])
         MPI_File_close(&fh);
 
         // convert the data in required format
-        for (int i = 0; i < localN; i++)
+        for (long long i = 0; i < localN; i++)
         {
-            for (int t = 0; t < nc; t++)
+            for (long long t = 0; t < nc; t++)
             {
                 localArr[t][i] = readArr[i * nc + t];
             }
@@ -323,7 +323,7 @@ int main(int argc, char *argv[])
         // Before starting the computation it is wise to exchange the ghost layers beforehand
         double *leftGhostLayer[nc], *rightGhostLayer[nc], *topGhostLayer[nc], *bottomGhostLayer[nc], *frontGhostLayer[nc], *backGhostLayer[nc];
 
-        for (int t = 0; t < nc; t++)
+        for (long long t = 0; t < nc; t++)
         {
             // sending the ghost layers
             if (ip != 0)
@@ -391,15 +391,15 @@ int main(int argc, char *argv[])
         }
 
         // Local and Global Minima Calculations
-        int sxi, syi, szi, isLocalMinima, isLocalMaxima;
-        for (int t = 0; t < nc; t++)
+        long long sxi, syi, szi, isLocalMinima, isLocalMaxima;
+        for (long long t = 0; t < nc; t++)
         {
             localMinima[t] = 0;
             localMaxima[t] = 0;
             globalMinima[t] = localArr[t][0];
             globalMaxima[t] = localArr[t][0];
 
-            for (int i = 0; i < localN; i++)
+            for (long long i = 0; i < localN; i++)
             {
                 globalMaxima[t] = (localArr[t][i] > globalMaxima[t]) ? localArr[t][i] : globalMaxima[t];
                 globalMinima[t] = (localArr[t][i] < globalMinima[t]) ? localArr[t][i] : globalMinima[t];
@@ -488,10 +488,10 @@ int main(int argc, char *argv[])
         }
 
         // Gathering the local and global extermas to a single process
-        for (int t = 0; t < nc; t++)
+        for (long long t = 0; t < nc; t++)
         {
-            MPI_Reduce(&localMinima[t], &totalLocalMinima[t], 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
-            MPI_Reduce(&localMaxima[t], &totalLocalMaxima[t], 1, MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_Reduce(&localMinima[t], &totalLocalMinima[t], 1, MPI_LONG_LONG, MPI_SUM, 0, MPI_COMM_WORLD);
+            MPI_Reduce(&localMaxima[t], &totalLocalMaxima[t], 1, MPI_LONG_LONG, MPI_SUM, 0, MPI_COMM_WORLD);
             MPI_Reduce(&globalMinima[t], &globalGlobalMinima[t], 1, MPI_DOUBLE, MPI_MIN, 0, MPI_COMM_WORLD);
             MPI_Reduce(&globalMaxima[t], &globalGlobalMaxima[t], 1, MPI_DOUBLE, MPI_MAX, 0, MPI_COMM_WORLD);
         }
@@ -523,16 +523,16 @@ int main(int argc, char *argv[])
         }
 
         // Printing number of local extremas
-        for (int t = 0; t < nc; t++)
+        for (long long t = 0; t < nc; t++)
         {
             if (t != 0)
                 fprintf(file, ", ");
-            fprintf(file, "(%d, %d)", totalLocalMinima[t], totalLocalMaxima[t]);
+            fprintf(file, "(%lld, %lld)", totalLocalMinima[t], totalLocalMaxima[t]);
         }
         fprintf(file, "\n");
 
         // Printing global extremas
-        for (int t = 0; t < nc; t++)
+        for (long long t = 0; t < nc; t++)
         {
             if (t != 0)
                 fprintf(file, ", ");
